Move helper to back to utils
