+++ packages/app/src/domain/international/infected-table-tile/components/wide-infected-table.tsx
@@ -9,7 +9,7 @@ import { InlineText } from '~/components/typography';
 import { Flag } from '~/domain/international/flag';
 import { useIntl } from '~/intl';
 import { asResponsiveArray } from '~/style/utils';
-import { getFilteredThresholdValues } from '~/components/choropleth/logic/get-filtered-threshold-values';
+import { getFilteredThresholdValues } from '~/utils/get-filtered-threshold-values';
 import { getMaximumNumberOfDecimals } from '~/utils/get-maximum-number-of-decimals';
 import { FilterArrayType } from '../infected-table-tile';
 import { MAX_COUNTRIES_START } from '../logic/common';
